requirements.txt
# DefectDetector Service Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image Processing
Pillow>=10.2.0
numpy>=1.26.0
opencv-python>=4.9.0

# ML Inference
ultralytics>=8.1.0
torch>=2.0.0

# Optional: ONNX Runtime (alternative to PyTorch)
# onnxruntime>=1.17.0

